Get rid of separate categorical_util module
