+++ src/glum_benchmarks/data/__init__.py
@@ -1,4 +1,3 @@
-from . import categorical_util
 from .create_housing import create_housing_raw_data, generate_housing_dataset
 from .create_insurance import (
     create_insurance_raw_data,
@@ -18,5 +17,4 @@ __all__ = [
     "simulate_glm_data",
     "generate_housing_dataset",
     "create_housing_raw_data",
-    "categorical_util",
 ]
